Flatten the Gravity qualifiers - Move all immediate qualifiers to the root of values - Fix old import path ( @cloudinary/base/values/gravity/qualifiers/focusOn/FocusOn ) to be ( @cloudinary/base/values/focusOn ) - The same was done for AutoFocus and Compass
